split render and img, add internal headers
