+++ include/engine/internal/img.h
@@ -1,0 +1,11 @@
+#ifndef IMG_INTERNAL_H
+#define IMG_INTERNAL_H
+
+#include "engine/img.h"
+#include "engine/internal/render.h"
+
+int init_sprites(void);
+void cleanup_sprites(void);
+void draw_all_sprites_and_gc(void);
+
+#endif /* IMG_INTERNAL_H */
