+++ include/engine/util.h
@@ -18,8 +18,6 @@
 
 #ifndef BIN_LOADER
 void quit_game(void); /* Request to quit the game */
-SDL_Window *get_window(void); /* Return the window */
-SDL_Renderer *get_renderer(void); /* Return the render context */
 #endif
 
 #endif /* UNTITLED_PLATFORMER_UTIL_H */
