+++ src/engine/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(amphora_engine
 	main.c
 	render.c
+	img.c
 	input.c
 	events.c
 	ttf.c)
