+++ src/game_loop.c
@@ -1,4 +1,5 @@
 #include "engine/game_loop.h"
+#include "engine/img.h"
 #include "engine/input.h"
 #include "engine/render.h"
 #include "engine/ttf.h"
@@ -13,7 +14,6 @@ AmphoraMessage *hello;
 AmphoraMessage *timer;
 AmphoraMessage *stationary;
 
-int player_idle_frames[NUM_PLAYER_IDLE_FRAMES] = { 1, 3 };
 int player_idle_idx = 0;
 bool walking = false;
 
@@ -21,10 +21,10 @@ void
 game_init(void) {
 	const char *welcome_message = "Hello, and welcome to the Amphora demo!";
 	const char *stationary_message = "I'm going to be fixed right here above the tower!";
-	SDL_Color font_color = { 0, 0, 0, 0xff };
+	SDL_Color font_color = { 0xff, 0xff, 0xff, 0xff };
 
-	init_sprite_slot(&p_char, player_idle_frames[player_idle_idx], 2, 4, 12, 120, false, 10);
-	init_sprite_slot(&building, 12, 4, 8, 96, 148, false, -1);
+	//init_sprite_slot(&p_char, player_idle_frames[player_idle_idx], 2, 4, 12, 120, false, 10);
+	//init_sprite_slot(&building, 12, 4, 8, 96, 148, false, -1);
 	create_string(&hello, Roboto, 32, 16, 16, font_color, welcome_message);
 	create_string(&timer, Merriweather, 32, -16, 16, font_color, "0");
 	create_stationary_string(&stationary, Merriweather, 16, 76, 132, font_color, stationary_message);
@@ -44,6 +44,7 @@ game_loop(Uint64 frame, const struct input_state_t *key_actions, SaveData *save_
 	if (key_actions->dash) {
 		p_movement_speed = 2;
 	}
+	/*
 	if (key_actions->left) {
 		p_char->flip = true;
 		p_char->x -= p_movement_speed;
@@ -54,15 +55,18 @@ game_loop(Uint64 frame, const struct input_state_t *key_actions, SaveData *save_
 		p_char->x += p_movement_speed;
 		walking = true;
 	}
+	*/
 	if (key_actions->quit) {
 		quit_game();
 	}
 
+	/*
 	if (frame - idle_anim > 30) {
 		if (++player_idle_idx >= NUM_PLAYER_IDLE_FRAMES) player_idle_idx = 0;
 		p_char->num = player_idle_frames[player_idle_idx];
 		idle_anim = frame;
 	}
+	*/
 
 	if (frame % 60 == 0) {
 		snprintf(timer_string, 128, "%d", SDL_atoi(timer_string) + 1);
@@ -77,10 +81,12 @@ game_loop(Uint64 frame, const struct input_state_t *key_actions, SaveData *save_
 	render_string(timer);
 	render_string(stationary);
 
+	/*
 	camera_location = get_sprite_center(p_char);
 	camera_location.x -= (get_resolution().x / 2);
 	camera_location.y = 0;
 	set_camera(camera_location.x, camera_location.y);
+	*/
 }
 
 void
